Fix another IPR slow edit reported by Eugene.
Accessing the blender_data user_map can be really
slow. We were checking this because we always
deleted all instances of an object that's been
updated. In the case of geo nodes, Blender only tell us that the object with the geo nodes has updated,
not the object that's being instance, so we were
always checking the user_map to figure out what object was being instanced, so we don't accidentally
delete the prototype.
Now, instead, we keep track of what instances are still around in a dictionary. After the loop through the depsgraph instances, we iterate through the dictionary; whatever instances that are not in the dictionary's list are ones that no longer exist and can
ben deleted.