For RMAN-23335
The previous fix for RMAN-23282 didn't fully
work. The pointer address apparently changes
with each parameter change, so it's not a good
solution to use if we want to do a search and
compare.

We now do something similar to what we do
when we do light soloing:
1) add a boolean to the output node that
indicates the shading network has a solo node
2) add another boolean to every shading node
to indicate whether or not it itself is the solo node.