More work towards RMAN-23591
This change helps with some usability.

* add a new static panel for the stylized looks
editor. This might be easier to use than the
floating editor panel. This required adding
a new stylized PropertyGroup to the scene.
* we now list all of the stylized filters in the
scene on the Filters tabs, not just the current
selected one. You can also remove, or move
the filters up and down the chain
* change the object search filter to just be a
regular Blender pointer property. This allows
the user to use the eye dropper to select an
object from the viewport to attach a PxrStylizedControl pattern.