For RMAN-23690
When we IPR to it, there seems to be minute
differences in the camera transform when we
call our update_view method, which seems to
trigger an edit, even if all you did was select geometry.
For now, don't call update_view since update_scene
will also handle camera transforms and edits.

Also, change the viewport text color to be RenderMan blue.