Some more speed ups for scene translation/exporting.
For numpy arrays, if we don't need to do any
element manipulation, we can simply pass the array's data property directly to the RtParamList, rather
than calling reshape() and then tolist() to convert
to a list. The combination of reshape() and tolist()
can be really slow.

With the other fixes, this cuts down one of Eugene's SIGGRAPH demo translation time from 1:06 to about
0:14.