Fix for RMAN-24005
It seems trying to check if the viewport is still in RENDERED mode,
is not suppported in a separate thread.
Why this crashes Blender when only
an undo happens is not clear.

For now, switch to using a Blender application timer, which should serve the same purpose as a thread for our needs.