For RMAN-24277
When an object is marked as updated, make sure to check its modifiers list for a geometry node tree.
If there is a tree, mark the object as also needing transform updating. Blender doesn't seem to mark it as a transform update, even though there may be object instancing nodes in the geometry node tree.