For RMAN-24315
Calling bpy.data.node_groups.new seems to crash Blender when PxrRamps are copied.

Note, the way we insert the ramp UI into PxrRamp  is that we have to create one of the native Blender ramp shading nodes, and then embed its ramp UI into our node. We created these native nodes in a hidden nodegroup.

The fix here is to rather than create a new node group for each PxrRamp, we'll just reuse the same node group for all the nodes that use ramps. The name of the native node is always stashed on the PxrRamp node so we can reference it later.

Finally, bumped the addon version to 27.1 and update the changelog.txt