Workaround fix for RMAN-24325
Reported on GitHub (#722). NaNs will appearing in the denoise result when using the Blender display driver.

For some reason there are NaNs in the variance channels when retrieving the buffers from the Blender display driver; this doesn't seem to happen when using OpenEXR.

For now, just use numpy.nan_to_num to remove any NaNs in the buffer before handing it to the denoiser.